"Correcciones y nueva migracion hecha"
> Aplique correcciones en modelos para corregir errores sobre la nueva migracion, agregue a gitignore archivos para evitar que se suban estos

+++ Models/ControlActividadesDTO.cs
@@ -96,7 +96,7 @@ namespace ControlActividades.Models
         public int MateriaId { get; set; }
         public DateTime FechaCreacion { get; set; }
         public DateTime FechaLimite { get; set; }
-        public int Puntaje { get; set; }
+        public decimal Puntaje { get; set; }
         public bool? Enviado { get; set; }
         public bool PermitirEntregasTarde { get; set; }
         public DateTime? FechaProgramada { get; set; }
@@ -132,7 +132,7 @@ namespace ControlActividades.Models
         public string NombreActividad { get; set; }
         public string Descripcion { get; set; }
         public DateTime FechaLimite { get; set; }
-        public int Puntaje { get; set; }
+        public decimal Puntaje { get; set; }
         public int MateriaId { get; set; }
         public bool? Enviado { get; set; }
         public DateTime? FechaProgramada { get; set; }
